Add helpers to update or remove query params and logic to implement on integrations page
