+++ src/scripts/helpers/helpers.js
@@ -52,17 +52,4 @@ const getCookieByName = (name) => {
     return value;
 }
 
-const getQueryParameterByName = (name, currentURL) => {
-    const queryParameterName = name.replace(/[`[\]]/g, '\\$&');
-    const url = !currentURL ? window.location.href : currentURL;
-    const regex = new RegExp(`[?&]${queryParameterName}(=([^&#]*)|&|#|$)`);
-    const results = regex.exec(url);
-
-    if (!results) return null;
-    if (!results[2]) return '';
-
-    return decodeURIComponent(results[2].replace(/\+/g, ' '));
-}
-
-
-export {updateMainContentAnchors, reloadWistiaVidScripts, gtag, getCookieByName, getQueryParameterByName}
+export {updateMainContentAnchors, reloadWistiaVidScripts, gtag, getCookieByName}
